set PHYCAS_ROOT then run refill_phycas_gui_app.sh
git-svn-id: https://phycas.svn.sourceforge.net/svnroot/phycas/trunk@636 b774380d-c90d-0410-997c-aec342e5d6fd

+++ installer/mac/iterm/makeDMG.sh
@@ -2,11 +2,11 @@
 NAME=Phycas
 VERSION=1.0
 MASTER_DMG=$NAME-$VERSION.dmg
-SOURCE_DIR="/Users/mholder/Documents/projects/iterm/iTerm/build/Deployment/"
 set -x
 files="PhycasGUI.app" 
 if test -d "$PHYCAS_ROOT"
 then
+	SOURCE_DIR="$PHYCAS_ROOT/installer/mac/iterm"
 	cd $PHYCAS_ROOT/installer/mac/diskimage/
 	TEMPLATE_DMG="template.dmg"
 	bunzip2 -k ${TEMPLATE_DMG}.bz2
@@ -24,11 +24,23 @@ then
 		rm -rf "$WC_DIR/$i"; \
 		ditto -rsrc "$SOURCE_DIR/$i" "$WC_DIR/$i"; \
 	done
-	WC_DEV=`hdiutil info | grep "$WC_DIR" | grep "Apple_HFS" | awk '{print $1}'` && hdiutil detach $WC_DEV -quiet -force
-	rm -f "$MASTER_DMG"
-	hdiutil convert "$WC_DMG" -quiet -format UDZO -imagekey zlib-level=9 -o "$@"
-	#rm -rf $WC_DIR
-	mv "$WC_DMG" "$MASTER_DMG"
+	WC_DEV=`hdiutil info | grep "$WC_DIR" | grep "Apple_HFS" | awk '{print $1}'` 
+	if test -z $WC_DEV
+	then
+		WC_DEV=`hdiutil info | grep "$WC_DIR" | grep '^/' | awk '{print $1}'` 
+	fi
+	if test -z $WC_DEV
+	then
+		echo "Could not unmount. Run 'hdiutil info' "
+		exit 2
+	else
+		hdiutil detach $WC_DEV -quiet -force
+		rm -f "$MASTER_DMG"
+		hdiutil convert "$WC_DMG" -quiet -format UDZO -imagekey zlib-level=9 -o "$@"
+		#rm -rf $WC_DIR
+		mv "$WC_DMG" "$MASTER_DMG"
+		mv "$MASTER_DMG" "$SOURCE_DIR/"
+	fi
 else
 	echo "PHYCAS_ROOT is not defined as a directory"
 	exit 1
